# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Meal prep subscription service API with weekly menus and stock management",
        "title": "PrepToPlate API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@preptoplate.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/admin/weekly-menus": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin only - Create a new weekly menu with meals and stock",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "weekly-menu"
                ],
                "summary": "Create weekly menu",
                "parameters": [
                    {
                        "description": "Menu data",
                        "name": "menu",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateWeeklyMenuRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.WeeklyMenu"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/weekly-menus/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin only - Get details of a specific weekly menu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "weekly-menu"
                ],
                "summary": "Get weekly menu by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.WeeklyMenu"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/weekly-menus/{id}/activate": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin only - Set a menu as active (deactivates all other menus)",
                "tags": [
                    "admin",
                    "weekly-menu"
                ],
                "summary": "Activate weekly menu",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/menu": {
            "get": {
                "description": "Get the currently active weekly menu with stock information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weekly-menu"
                ],
                "summary": "Get active weekly menu",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.WeeklyMenu"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateWeeklyMenuRequest": {
            "type": "object",
            "required": [
                "meals",
                "week_start_date"
            ],
            "properties": {
                "meals": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/models.MenuMealInput"
                    }
                },
                "week_start_date": {
                    "type": "string"
                }
            }
        },
...
        "models.Meal": {
            "type": "object",
            "properties": {
                "calories": {
                    "type": "integer"
                },
                "carbs": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "fat": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "description": "stored in cents",
                    "type": "integer"
                },
                "protein": {
                    "type": "integer"
                }
            }
        },
        "models.MenuMealInput": {
            "type": "object",
            "required": [
                "meal_id",
                "stock"
            ],
            "properties": {
                "meal_id": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "models.WeeklyMenu": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "meals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.WeeklyMenuMeal"
                    }
                },
                "week_start_date": {
                    "type": "string"
                }
            }
        },
        "models.WeeklyMenuMeal": {
            "type": "object",
            "properties": {
                "available_stock": {
                    "type": "integer"
                },
                "initial_stock": {
                    "type": "integer"
                },
                "meal": {
                    "$ref": "#/definitions/models.Meal"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}